Remove duplicated frontend tests
